send_telegram_test.py
#!/usr/bin/env python3
"""
Test script to send implementation plan to Telegram bot
"""
import requests
import os
import sys

# Load environment variables
try:
    from dotenv import load_dotenv
    load_dotenv('.env')
except ImportError:
    pass

BOT_TOKEN = os.getenv('TELEGRAM_BOT_TOKEN')
CHAT_ID = os.getenv('TELEGRAM_CHAT_ID')

if not BOT_TOKEN or not CHAT_ID:
    print("Error: TELEGRAM_BOT_TOKEN or TELEGRAM_CHAT_ID not found in environment")
    print("Please set these in your .env file")
    sys.exit(1)

IMPLEMENTATION_PLAN = """
🎉 === TELEGRAM ARCHITECTURE & METRICS ENHANCEMENT ===

📋 IMPLEMENTATION COMPLETE

✅ Files Created:
- telegram_system_monitor.py (~400 lines)
  - SystemArchitectureDisplay class
  - SystemMetricsCollector class
  - SystemMetricsCache class

✅ Files Modified:
- telegram_dashboard.py (~200 lines added)
- telegram_bot_handler.py (~10 lines added)
- requirements.txt (psutil added)

✅ New Telegram Commands:

1. /architecture
   - Show system architecture summary with interactive buttons
   - Options: View Full Diagram, Component Details, Data Flow

2. /sysmetrics
   - Quick summary of all system metrics (30s cached)
   - Detailed drill-down for: CPU, Memory, Disk, Network, Database, API, ML, Security, Trading

🏗️ SYSTEM ARCHITECTURE VIEW

The /architecture command displays:
- High-level 3-layer architecture
- Trading Engine (VNPy-based)
- AI/ML Services (LSTM/Transformer)
- Security Framework (JWT/Trust/Audit)
- Infrastructure (PostgreSQL, Redis, Prometheus, Grafana)

📊 SYSTEM METRICS VIEW

The /sysmetrics command displays:
- CPU: Usage%, cores, load average, per-core breakdown
- Memory: Total, used, available, swap usage
- Disk: Usage per mount point
- Network: Traffic (sent/recv), packets, errors, connections
- Database: PostgreSQL connections, Redis memory
- API: Health status, uptime
- ML: Model availability status
- Security: Trust scores, auth stats
- Trading: Positions, P&L, engine status

🚀 USAGE EXAMPLES:

/architecture              # Show summary
/architecture full         # Full diagram
/architecture details ai_ml_services  # Component details
/architecture flow         # Data flow

/sysmetrics               # Quick summary (cached)
/sysmetrics cpu           # Detailed CPU metrics
/sysmetrics memory        # Detailed memory metrics
/sysmetrics db            # Database metrics
/sysmetrics trading       # Trading metrics

🔧 FEATURES:

✓ Hybrid caching: 30s TTL for summary, real-time for details
✓ Thread-safe cache implementation
✓ Error handling with graceful degradation
✓ Interactive inline keyboards for navigation
✓ Mobile-friendly formatting (message < 4096 chars)
✓ Security filtering (no sensitive data exposed)
✓ Rate limiting support
✓ psutil integration for system resources

📦 DEPENDENCIES:

- psutil>=5.9.0 (installed ✅)

🎯 READY TO USE!

Try these commands now:
/architecture
/sysmetrics

For full help: /help

Generated: {}

""".format(__import__('datetime').datetime.now().strftime('%Y-%m-%d %H:%M:%S'))

# Send message to Telegram
api_url = f"https://api.telegram.org/bot{BOT_TOKEN}/sendMessage"

payload = {
    'chat_id': CHAT_ID,
    'text': IMPLEMENTATION_PLAN,
    'parse_mode': 'HTML'
}

response = requests.post(api_url, json=payload, timeout=30)

if response.status_code == 200:
    result = response.json()
    print("✅ Message sent successfully!")
    print(f"Message ID: {result['result']['message_id']}")
    print("\n📱 Check your Telegram bot to see the implementation plan!")
else:
    print(f"❌ Failed to send message: {response.status_code}")
    print(response.text)
    sys.exit(1)